Bump version to $VERSION

+++ src/common/controller/remove.js
@@ -4,7 +4,9 @@ const constants = require('../helpers/constants')
 
 exports.removeMany = async (req,res, model) => {
   const ids = req.body.ids;
-  IsArray(ids,res);
+  if (!ids || !Array.isArray(ids) || ids.length === 0) {
+    return Response(res, 400, "Not Found Ids");
+  }
   try {
     await model.deleteMany({ _id: { $in: ids } });
     Response(res, 200, constants.DELETE_SUCCESS);
